# faster access to Context and ContextPath