This PR is a small part of #64 that helps to access the current context and the context without the need to look at the corresponding definition. These variables are intensively used each time that the context of a symbol needs to be resolved.