@rocky 2021-11-25
If we could get a benchmark on this, I'd appreciate it. 

On my list of things to do is benchmark all of the changes since 4.0.0 to get a sense of what improvements there have been.