@mmatera 2021-11-25
Here are the benchmarks. 

From running `timeit ` in jupyter

| expr          | master [us] |  faster_context [us] | Improvement |
| ------------  | ------- | ---------------  |   ----- | 
|  `a`          |   12.7  |       9.11       | $28\%$ |
|  `E`          |   17.6  |       13.1       | $25\%$|
|  `F[]`        |   42.6  |       38.6       | $9.4\%$|
|  `2*b`        |   156  |       150       |   $3.8\%$|
|  `Plus[a,b]`  |   157  |       146       | $7\%$|
|  `a+1+a+1...` |   821  |       779        |$5\%$|


Code:
```
from mathics.session import MathicsSession
from mathics.core.symbols import Symbol
from mathics.core.expression import Expression
from mathics.core.atoms import Integer, Real


session = MathicsSession()
evaluation = session.evaluation

print("evaluate an undefined symbol")
%timeit session.evaluate("a")

print("evaluate a symbol")
%timeit session.evaluate("E")

print("evaluate an expression")
%timeit session.evaluate("F[]")

print("evaluate a simple product")
%timeit session.evaluate("2*b")

print("evaluate a simple addition")
%timeit session.evaluate("Plus[a,b]")

print("evaluate a longer addition")
%timeit session.evaluate("a+1+a+1+a+1+a+1+a+1+a+1+a+1+a+1+a")
```

And these are some results from mathics-benchmarks:

![imagen](https://user-images.githubusercontent.com/2862288/143487527-5e2ac83b-12f9-4923-827d-fa9f00501b10.png)
![imagen](https://user-images.githubusercontent.com/2862288/143487537-8c3b7e74-d809-4df0-867a-e13fc1023429.png)
![imagen](https://user-images.githubusercontent.com/2862288/143487573-afa742e5-a4c7-4dd8-a948-8ad3e9c58831.png)
![imagen](https://user-images.githubusercontent.com/2862288/143487599-54d5d448-6f3e-4329-8c20-ba5825ee3c39.png)