@mmatera 2021-11-25
Again, I found funny some of the results in mathics-benchmarks: `Integer[1]` seems to be 14% slower than in master, but I found it hard to imagine where this overhead is coming taking into account the implemented changes...