@rocky 2021-11-26
> Again, I found funny some of the results in mathics-benchmarks: `Integer[1]` seems to be 14% slower than in master, but I found it hard to imagine where this overhead is coming taking into account the implemented changes...

I am not seeing this. In fact the the "faster_context" branch is faster.

Here is what I did. I `git clone`d  the master and `faster_contex`t branches in separate virtual environments.

Then I went into ipython in each and rean the setup code listed above

```
from mathics.session import MathicsSession
from mathics.core.symbols import Symbol
...
```

Here is what I see. First `master`:

```
In [9]: %timeit session.evaluate("Integer[1]")
%timeit session.evaluate("Integer[1]"))
139 µs ± 330 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [10]: %timeit session.evaluate("Integer[1]")
%timeit session.evaluate("Integer[1]"))
140 µs ± 431 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

Next `faster_context`

```
In [15]: %timeit session.evaluate("Integer[1]")
%timeit session.evaluate("Integer[1]"))
134 µs ± 517 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [16]: %timeit session.evaluate("Integer[1]")
135 µs ± 734 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```