@rocky 2021-11-26
I was able to see the speedups reported in lookup up single simples (in a loop of 1000 times, so this is kind of expected). 

The greatest mathics-benchmark difference I am not seeing though. Something is weird with mathics benchmark I think.