@rocky 2021-11-26
In resolving the CHANGES conflicts, I made some of the simpler typographical changes. However deeper changes should be made in a separate PR.